skyroute/backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-socketio==5.11.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.6
numpy==1.26.3
scipy==1.11.4
networkx==3.2.1
shapely==2.0.2
httpx==0.26.1
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.22
